requirements.txt
Babel==2.5.3
certifi==2018.1.18
chardet==3.0.4
click==6.7
dominate==2.3.1
Flask==0.12.2
Flask-Babel==0.11.2
Flask-Bootstrap==3.3.7.1
flask-nav==0.6
Flask-SQLAlchemy==2.3.2
Flask-Table==0.5.0
Flask-WTF==0.14.2
grpcio==1.10.0
grpcio-tools==1.10.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
netaddr==0.7.19
protobuf==3.5.2
protobuf3-to-dict==0.1.5
py-radix==0.10.0
pytz==2018.3
requests==2.18.4
six==1.11.0
socketIO-client==0.7.2
SQLAlchemy==1.2.5
urllib3==1.22
visitor==0.1.3
websocket-client==0.47.0
Werkzeug==0.14.1
WTForms==2.1
